File re-structuring, fixed the testing function calls, and fixed the update_honeypot_session issue
